Support user-provided TLS certificates for Antrea UI HTTPS
This enables cert-manager support.
To facilitate this, we replace the 2 secrets (one for the private key
and one for the cert) with a single secret of type
kubernetes.io/tls). When using Helm to generate the secret, we use name
"antrea-ui-tls-helm". When the user / cert-manager provides the secret,
the default name is "antrea-ui-tls". We don't use the same name to avoid
possible conflicts when switching between http "methods" in the chart.

We also add support for automatic reloading of the nginx configuration
if the certificate changes. This simplifies the process of certificate
rotation.

Signed-off-by: Antonin Bas <abas@vmware.com>