Add new reference deployment to documentation
Ingress with Nginx + cert-manager (Let's Encrypt) on EKS

Production users are very likely to use an Ingress Controller to
terminate TLS for the Antrea UI, so we document this case.

We also update the getting-started documnent to mention TLS termination
at the Ingress, and to mention the "secret" HTTPS method when TLS
termination happens at the UI server.

Signed-off-by: Antonin Bas <abas@vmware.com>