Ability to proxy arbitrary K8s APIs (#21)
We add a new API endpoint which proxies GET requests to the K8s
api-server. This removes the need to define custom APIs for each use
case that requires directly retrieving data from the K8s API. At the
moment we can use this API for the web client "Summary" page, which
displays information from the AntreaAgentInfo and AntreaControllerInfo
CRs. As a result, the specialized "/info" endpoint is now deprecated,
and will be removed in the near future (July 2023).

The list of API resources being proxied is determined by:

1) a global variable in the backend code; if a client attempts to access
   an API that is not being proxied, the "Not Found" status code will be
   returned. Same if the HTTP verb being used is not GET.
2) K8s RBAC permissions for the antrea-ui backend; the list / get verbs
   must be granted for the resources being proxied.

We use the Go standard library reverse proxy implementation. Note that
the implementation was greatly improved in Go 1.20, but we currently use
Go 1.19 for all Antrea projects.

Fixes #10

Signed-off-by: Antonin Bas <abas@vmware.com>