Improve the cookieSecure config parameter (#31)
We expose it in the Helm chart as security.cookieSecure. The default
behavior when installing the Helm chart does not change: if HTTPS is
enabled, it is set to true, otherwise it is set to false. However, by
exposing it, we enable users who terminate TLS at Ingress (and hence
keep HTTPS disabled in Antrea UI) to enable that option, which hardens
security.

We also enable it by default in the Viper config, which does not really
have an impact on users.

Note that it is not possible for a non-secure origin (HTTP website) to
add the Secure attribute. Most modern browsers will not allow it and the
cookie will be discarded. This is why we do not unconditionally set the
Secure attribute. This restriction does not apply for localhost.

Signed-off-by: Antonin Bas <abas@vmware.com>