Move auth routes out of API
I believe it makes more sense to decouple authentication from the rest
of the API. The API only requires a JWT token, which in theory could be
obtained through different means.
With this change, all the cookie manipulation logic is moved out of the
API, which makes the API itself more "RESTful".

Signed-off-by: Antonin Bas <abas@vmware.com>