Support OIDC authentication for SSO (#38)
Antrea UI can now be deployed with SSO support, by connecting to
OIDC-conformant identity providers.

There are 2 main methods to enable OIDC support in Antrea UI:

1. Use the built-in Dex support to connect to an identity provider of
   your choice.
2. Configure Antrea UI to connect to an OIDC-conformant provider. The
   provider could be an external instance of Dex managed by you, an OIDC
   service such as Okta, Auth0, etc.

When enabling OIDC, Basic authentication using the admin password can be
disabled.

The OIDC client is part of the Antrea UI backend (using the "Backend For
Frontend" authentication pattern). This has the following advantages:

* we can usse the OAuth2 "Authorization Code Flow", as we can store an
  application secret securely in the backend
* no need for PKCE
* the frontend stays mostly the same and we don't really need to
  differentiate between authentication mechanisms

The design for API authorization does not change. When the user is
authenticated using OIDC, the backend issues a short-lived refresh token
(30 minutes) that is stored in a session cookie. The refresh token is
used to obtain JWT access tokens which grant access to the API.

Fixes #28

Signed-off-by: Antonin Bas <abas@vmware.com>