Fix chart for older Helm versions
When using Helm v3.6, installing the chart fails with:
`error calling eq: incompatible types for comparison`.

This is because in older versions of Helm (built with older versions of
Go), the `eq` built-in function does not support comparing a `nil` value
to a non-`nil` value.

We can workaround this issue by converting the boolean value to a string
with `toString`.

Fixes #42

Signed-off-by: Antonin Bas <abas@vmware.com>