Add support from fetching FeatureGates information
The information is retrieved by the API server by connecting to the
Antrea Service (provided by the Antrea Controller). If the server is
running outside of the K8s cluster (dev environment), we automatically
set up port forwarding to forward from localhost to the Antrea
Controller Pod.

The FeatureGates information is displayed on the Summary page of the UI.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>