Use v1 tag for benc-uk/workflow-dispatch Github Action (#348)
It seems that tag v121 no longer exists. I do not recall why v1 was not
used in the first place, but it seems that it is the right thing to do
now. antrea-io/antrea also uses v1. This tag will make sure that the
latest available release for major version v1 is used, every time the
action runs.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>