Bump axios to 1.7.5 (#468)
The base URL for relative URLs no longer defaults to http://localhost
but now defaults to the "origin" (window.location.href). This led to
some Node test failures. There are several ways to address this, but we
chose to simply use the origin as the URL for the nock test server.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>