Update to ESLint v9
As per https://eslint.org/docs/latest/use/migrate-to-9.0.0. The main
change is that the config format has changed.

We also update some configuration files to match the latest version of
the relevant Vite template:
https://github.com/vitejs/vite/tree/main/packages/create-vite/template-react-ts

As part of this, we enable type checking when running "yarn build" (it
was disabled until now, and "vite build" only performs transpilation),
and address type checking errors as necessary.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>